requirements.txt
astroid==2.12.10
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.5.1
elastic-transport==8.4.0
elasticsearch==8.4.2
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
psutil==5.9.2
pylint==2.15.3
PyYAML==6.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.5
urllib3==1.26.12
wrapt==1.14.1
psutil==5.9.2
getmac==0.8.3